+++ .github/workflows/e2e-multiple-k8s-clusters.yaml
@@ -23,7 +23,6 @@ jobs:
       fail-fast: false
       matrix:
         package:
-          - replicationfailure
           - changetostandalone
           - changetoprimary
           - changetosecondary
@@ -47,6 +46,9 @@ jobs:
           - package: replication
             backup-transfer-part-size: 3Mi # smaller than the PVC size
             ginkgo-flags: --label-filter="misc"
+          - package: replication
+            backup-transfer-part-size: 3Mi # smaller than the PVC size
+            ginkgo-flags: --label-filter="backup-failure"
     #runs-on: "ubuntu-22.04"
     runs-on: mantle_large_runner_16core
     timeout-minutes: 120
